replaced all __init__.py classes to be easier to import
removed potentials problem with import cycle
removed class definitions from __init__.py to put it in core.py
cleaned import from __init__.py to import directly from declarants modules.

added a special SymbolTableBuilder for constants.
