+++ src/booleano/parser/trees.py
@@ -53,7 +53,7 @@ class ParseTree(object):
         """
 
         :param root_node: The root node of the parse tree.
-        :type root_node: :class:`booleano.operations.OperationNode`
+        :type root_node: :class:`booleano.operations.core.OperationNode`
 
         """
         self.root_node = root_node
@@ -99,7 +99,7 @@ class EvaluableParseTree(ParseTree):
         """
 
         :param root_node: The root node of the parse tree.
-        :type root_node: :class:`booleano.operations.OperationNode`
+        :type root_node: :class:`booleano.operations.core.OperationNode`
         :raises booleano.exc.InvalidOperationError: If the ``root_node`` is an
             operand that doesn't support logical values.
