+++ tests/parsing/test_parsers.py
@@ -34,13 +34,13 @@ from __future__ import unicode_literals
 from nose.tools import eq_, assert_raises
 import six
 
-from booleano.operations.variables import BooleanVariable
-from booleano.parser import Grammar, ConvertibleParser, EvaluableParser
+from booleano.parser.grammar import Grammar
+from booleano.parser.parsers import EvaluableParser, ConvertibleParser
 from booleano.parser.scope import Namespace
 from booleano.parser.parsers import Parser
 from booleano.operations import (Not, And, Or, Xor, Equal, NotEqual, LessThan,
     GreaterThan, LessEqual, GreaterEqual, BelongsTo, IsSubset, String, Number,
-    Set, Variable, Function, PlaceholderVariable, PlaceholderFunction)
+    Set, PlaceholderVariable, PlaceholderFunction)
 from booleano.parser.testutils import BaseGrammarTest
 from booleano.exc import ScopeError, BadExpressionError
